fix(profile): move allowlist out of 'use server' file (BUGS-12)
Next.js 16+ rejects non-async-function exports from `'use server'` files at
action-invocation time, which 500s the entire profile wizard whenever any
form is submitted. The check is runtime-only — the build does NOT catch it,
so the regression shipped green to dev.checklyra.com and broke every wizard
step until discovered by user testing on 2026-05-04.

Move ALLOWED_PROFILE_FIELDS, AllowedProfileField, and isAllowedProfileField
to a sibling profile-fields.ts module. The runtime allowlist enforcement in
updateProfileFields is unchanged (CodeQL alert #2 / CWE-250 protection
preserved). Test imports the const from the new module.

Add scripts/check-server-action-exports.sh to pr-checks.yml so any future
non-async-function export in a `'use server'` file fails CI rather than
shipping silently.
