+++ CLAUDE.md
@@ -196,7 +196,9 @@ These have caused real bugs. Read before making related changes:
 
 17. **edit_block corrupts markdown on long/complex content**: Desktop Commander's `edit_block` tool intermittently corrupts `.md` files when the new content is large (~1KB+) and contains em-dashes, multiple backticks, and tildes. Two observed symptoms: (a) tildes get escaped to `\~`, (b) wholesale deletion of unrelated lines. Discovered 2026-04-29 during BUGS-4 work; tracked under BUGS-5. Workarounds: (1) for multi-change markdown updates, prefer manual paste in editor (bypasses every tool layer); (2) for surgical edits, use `python3 -c 'with open(...) as f: ... f.write(...)'` via bash (direct filesystem write); (3) ALWAYS run `git diff <file>` after any markdown edit and revert if corruption seen. `edit_block` is fine for code files — only unreliable on `.md`.
 
-18. **Cloudflare Workers Builds posts a check on every PR by default**: The `lyra-maintenance` Workers Builds Git integration triggers on every push to any branch and posts a "Workers Builds: lyra-maintenance" check to GitHub. The build fails on PRs that don't change worker code (no-op build) and — even though the check is NOT in branch-protection required-checks — its failure status blocks GitHub auto-merge, forcing admin-merge. Fix: in Cloudflare dashboard → Workers & Pages → `lyra-maintenance` → Settings → Build → Build watch paths, set Include paths to `wrangler.toml, scripts/lyra-maintenance-worker.js`. With watch paths set, pushes that don't touch those files skip the build entirely and no GitHub check is posted. Caveats from the docs: watch-path matching is bypassed (i.e. build always runs) for empty pushes, pushes with 3000+ file changes, or pushes with 20+ commits — so very large bulk merges will still trigger a no-op build. Discovered 2026-05-04 from PR #98. **Verified post-fix behaviour (KAN-174, 2026-05-04)**: on a doc-only commit (5d5ee1f, CLAUDE.md only) the check still posts but with `conclusion: success` and `started_at == completed_at` (zero duration — build skipped at source). Auto-merge therefore not blocked. The fix changes the check from FAILURE → SUCCESS rather than removing it entirely.
+18. **`'use server'` files can ONLY export async functions**: In Next.js 16+ / React 19, any non-async-function export from a file declared `'use server'` (e.g. `export const FOO = [...]`, `export class X`, `export function syncFn()`) is rejected at *action-invocation time* with `Error: A "use server" file can only export async functions, found "X"`. The build does NOT catch this — it only fires when the action module is loaded by an action call in production, so the bug ships green to the dev preview and 500s every form submission on the affected route. Discovered 2026-05-04 from a regression where `ALLOWED_PROFILE_FIELDS` was added to `src/app/dashboard/profile/actions.ts` for testability — this broke every step of the profile wizard on dev. Tracked under BUGS-12. **Fix pattern:** move constants, types, and type guards to a sibling `.ts` module (e.g. `profile-fields.ts`) and import them from the action file. `export type` is fine (types are erased), but anything with a runtime shape must live elsewhere. **Guard:** `scripts/check-server-action-exports.sh` (run from `pr-checks.yml`) statically detects this pattern and fails any PR that re-introduces it. Allow-list with `// server-action-exports-ok: <reason>` only with explicit justification.
+
+19. **Cloudflare Workers Builds posts a check on every PR by default**: The `lyra-maintenance` Workers Builds Git integration triggers on every push to any branch and posts a "Workers Builds: lyra-maintenance" check to GitHub. The build fails on PRs that don't change worker code (no-op build) and — even though the check is NOT in branch-protection required-checks — its failure status blocks GitHub auto-merge, forcing admin-merge. Fix: in Cloudflare dashboard → Workers & Pages → `lyra-maintenance` → Settings → Build → Build watch paths, set Include paths to `wrangler.toml, scripts/lyra-maintenance-worker.js`. With watch paths set, pushes that don't touch those files skip the build entirely and no GitHub check is posted. Caveats from the docs: watch-path matching is bypassed (i.e. build always runs) for empty pushes, pushes with 3000+ file changes, or pushes with 20+ commits — so very large bulk merges will still trigger a no-op build. Discovered 2026-05-04 from PR #98. **Verified post-fix behaviour (KAN-174, 2026-05-04)**: on a doc-only commit (5d5ee1f, CLAUDE.md only) the check still posts but with `conclusion: success` and `started_at == completed_at` (zero duration — build skipped at source). Auto-merge therefore not blocked. The fix changes the check from FAILURE → SUCCESS rather than removing it entirely.
 
 ## Supabase Migration Rules
